using comparison endpoint only when a comparison is enabled
